test: align generated NIP specs and relay list test suite
- refresh generated feature snapshots and test project references
- update relay-list and whitelist validator tests for recent behavior changes
- adjust test host wiring for current auth and event handler composition

+++ test/Netstr.Tests/AuthTests.cs
@@ -56,7 +56,7 @@ namespace Netstr.Tests
                     ["relay", "ws://localhost"],
                     ["challenge", auth[1].ToString()]
                 ],
-                Kind = EventKind.Auth
+                Kind = (long)EventKind.Auth
             };
 
             e = Helpers.FinalizeEvent(e, Alice.PrivateKey);
@@ -97,7 +97,7 @@ namespace Netstr.Tests
                     ["relay", "ws://localhost"],
                     ["challenge", auth[1].ToString()]
                 ],
-                Kind = EventKind.Auth + 1
+                Kind = (long)EventKind.Auth + 1
             };
 
             e = Helpers.FinalizeEvent(e, Alice.PrivateKey);
